Use ,omitempty in DNSRecord type
Using ,omitempty means empty fields won't be added when marshalling to
JSON. This is what we need when creating a new DNS record so that we
don't try to send some of the read-only fields returned by the API and
means we don't need separate types for new records and responses.

Also split the response types out: Listing records returns an array and
creating or updating does not.