Refactor method signatures to use zone ID
This is a breaking change!

This solves inconsistencies in the API where we either took a zone ID, zone
name or Zone struct as input to the library methods. In #18 we agreed on
taking a zone ID (string) as input since it's the lowest common denominator.

I also took the liberty of aligning the preparation for calling makeRequest()
inside the functions, meaning we create a `uri` variable with the endpoint to
be called and pass that to makeRequest() instead of specifying it directly in
the method call.

There is also some minor refactors related to renaming `id` to `rid` in the
methods related to DNS, in order to avoid confusion with the new `zid` var.