zone: Implement zone edit methods
Add four new methods: EditZone, ZoneSetPaused, ZoneSetVanityNS and
ZoneSetPlan.

This also sets most of the ZonePlan fields to `omitempty` so that we can
pass a plan ID when changing a zone's plan.

EditZone wouldn't normally be called directly but is made public in case
the consumer chooses to do so.

ZoneSetPaused currently can only pause and not un-pause. This needs to
be fixed.