Handle API backend failures with a better error message (#83)
Everyone has bad days, CloudFlare is not immune to that, however
currently when CloudFlare's own API has a bad day, currently a
very verbose error message is returned that can cause very
large log lines.

To address this we catch HTTP 522,523,524, and 504 with a
"Service Failure" message and do not include the responce
body back with it, since there is no need to return it
as it will most likely not be useful to anyone.