feat(DNSRecords): return all records
Until pagination is implemented, the current implementation of
DNSRecords is lacklustre as it only returns a limited number of records
(20 by default).

This change updates DNSRecords to always return all records to the
caller, allowing it to use the data as it seems fit. It fetches the
maximum number of records per-page as allowed by the API to minimise the
API calls needed.