feat(auth): implement User Service authentication
This commit adds constants for choosing an authentication type
(key+email or user-service key) and a method for setting the
authentication type. It also updates `makeRequest` to set the
appropriate headers for authentication.

Some new endpoints require X-Auth-User-Service-Key instead of X-Auth-Key
+ X-Auth-Email.

This change makes no assumption that only one will be used at a time;
it's perfectly possible to tell the API to send both authentication
types (by using a bitmask), however no guarantee is made this this will
do what you expect.