Return response when page rule created
This commit is a (breaking) change to the `CreatePageRule` signature,
making it consistent with other `Create~` functions:

    func (api *API) Create~(zoneID string, ~) (*~Response, error)

Aside from motivations of consistency, without this API the user is
unable to discover the server-created ID for the page rule without
enumerating `ListPageRule` before and after creation - and assuming they
are the only user at that instant.

C.f. discussion in #108