Refactor API to accept an interface for the HTTP client
It is possible to provide an alternative client other than the
http.DefaultClient, however it is not possible to use another
implementation. An alternative implementation may be useful for
OpenTracing or a circuit breaker, for example.

Use an interface abstracting the implementation for the HTTP client,
rather than depending on a specific implementation.

The `HTTPRequester` name isn't ideal, but `HTTPClient` is already taken
by the options function.