feat(zones): allow pagination of zones
This changes adds a new method to the API, allowing for the pagination
of zones using the functional options pattern. In addition, it allows
the client to pass context through to the underlying HTTP library.

A functional option has been provide to configure the request's
current pagination, using the pre-existing PaginationOptions struct.

The high-level `ZoneIdByName` method has been modified to utilize this
new method.

Depends-On: #219
Bug: #182