Add support for validating expressions
This creates support for validating arbitary expressions. One caveat
here is that I've neeed to update the HTTP handler since the API
response is a HTTP 400 and results in the default case statement being
used. This stops the execution as it is treated as a HTTP error.
Instead, this adds a new condition for it that just returns the usuable
JSON response.