Swap `CustomPage.URL` to `interface`
When the custom page is in a default state (that is, Cloudflare default
pages) the API response has the URL set to `null` which breaks the
existing struct as it's expecting a string. To handle this case and
allow us to revert to the default values, we need to update the
`CustomPage.URL` to be an `interface{}` which will account for all
possible values here.