Add support for managing Custom Pages (#240)
* Rename to custom_pages to be more specific

* Add support for fetching custom pages

This introduces the ability to fetch the zone or account level custom
pages based on either the `ZoneID` or `AccountID` respectively.

To control which to fetch, you need to provide `CustomPageOptions` which
either takes a `ZoneID` or `AccountID` value.

Zone API reference: https://api.cloudflare.com/#custom-pages-for-a-zone-list-available-custom-pages
Account API reference: https://api.cloudflare.com/#custom-pages-account--list-custom-pages

* Add support for updating custom pages

* Add ability to fetch a single type of custom page

* Swap `CustomPage.URL` to `interface`

When the custom page is in a default state (that is, Cloudflare default
pages) the API response has the URL set to `null` which breaks the
existing struct as it's expecting a string. To handle this case and
allow us to revert to the default values, we need to update the
`CustomPage.URL` to be an `interface{}` which will account for all
possible values here.