Make `ZoneIDByName` safer to use for ambiguous zone names (#236)
Updates the `ZoneIDByName` function to be more intuitive (and safe) in
the event that multiple zones are returned by the API. It does this by
ensuring that if we have a response that includes more than a single
zone, an account ID *must be* defined to distinguish which account we
are intending to operate on. This enforcement then feeds into the
following step where the filtering not only matches on `zone.Name` but
the account ID provided as well. Lookups where the zone is unique will
continue to operate as normal.

Included in this change is updating `Account.Settings` to be a pointer
within the struct to enable the `Zone` struct to omit it should it not
be defined.

Fixes #234