Make rate limiting `correlate` optional (#281)
While looking into an issue in the Terraform provider, I've found that
the `correlate` values for rate limiting are always being sent
regardless of whether the value is set or not. This isn't really a big
deal as I'm assuming the API endpoint just disregard it but for the
Terraform provider, it does make the state management more difficult and
messy than what it needs to be due to the nested types.

Instead of always sending the value, I've updated the struct to allow
`nil` and to omit when the values are empty resulting in a fix for the
issues downstream.