zone: Swap to using subscription API for setting plans
While debugging an issue in the Terraform provider[1] I've come across
a broken endpoint whereby attempting to set a zone's rate plan
silently fails using this method. After a chat with some Cloudflare
Engineers, we're going to move to using the Subscription API instead
which has this functionality and is the intended approach to updating
a zone's rate plan.

[1]: terraform-providers/terraform-provider-cloudflare#488