Fix API token support for create/update/list worker routes
Previously, creating or updating a worker route would default to the deprecated `filters` endpoints when Enabled == false and ScriptName == "". This would cause errors when using API tokens since the filter endpoints don't support API tokens. This change will use the `routes` endpoints in that scenario instead, which do support API tokens.

It also updates `ListWorkerRoutes` to use the `routes` endpoint instead of `filters` when an API token is supplied.

Deleting worker routes suffers from the same problem, but that will be fixed in #367