access_identity_provider: Consolidate configuration fields
While looking integration Access Identity Providers into Terraform, I
found that the configuration fields are super awkward to work with as we
either:

- Disregard Go's type system and make everything an `interface`; or
- Restrict the ability to use configuration options in a single step as
  we don't know what Type the incoming data actually is.

To rectify this, I've consolidated all the fields into a single struct
and move the onus to the end user to correctly define the fields that
are required. A nice side effect of this is that the fields aren't
repeated for each configuration and reduces the toil if/when a field is
updated.