Update `*ListResponse` structs to match API responses
When the work for
terraform-providers/terraform-provider-cloudflare#597 was started,
the `messages` object was either empty array or a single string (per
the original `cloudflare.Response`[1] struct) but the API response is
now spitting out an array of `message` objects inside of the current
array.

To fix the issue, the two structs have been updated to conform to the
required objects.

[1]: https://github.com/cloudflare/cloudflare-go/blob/master/cloudflare.go#L363