Remove duplicated Access Policy structs (#432)
When I was first putting Access support together, I made a boo-boo and
duplicated the Access Group structs into the Access Policy itself. This
shouldn't have been the case as they are the same structs, but they
should be tied to the groups and only _referenced_ in the policies
instead.