logpush: Handle invalid destination configuration
In the event that the destination configuration of a logpush job
challenge returns an erorr, we don't currently catch that. Instead, we
return the ownership challenge struct successfully which is slightly
misleading.

This introduces more error checking on the API itself to inspect the
response payload for the `valid` parameter being set to true which
handles the scenarios where ownership validation fails.

Fixes #499