argo_tunnel: make V4 API response compatible
While diagnosing some issues implementing the `cloudflare_argo_tunnel`
resource in Terraform, I noticed that the API response was lacking the
outer response structure that the other V4 endpoints have.

The early implementation of Argo Tunnel endpoints didn't conform to the
V4 API standard response structure. This has been remedied going forward
however to support older clients this isn't yet the default. An explicit
`Accept` header is used to get the V4 compatible version.

Along with this change, is removing the use of the `context` package in
favour of existing request helpers. I opted not to reinvent a specific
case for this right now because:

- It's highly unlikely these requests will need to be cancelled; and
- This is currently a snowflake scenario and I didn't want to build a
  new interface after only needing it once. I'd prefer to wait for more
  use cases before expanding it.

As this change is internal, it won't impact clients and we can revert
this once the default response is the V4 compatible version without
interruption.