Better error handling for API errors
This introduces a change to how we raise and interact with exceptions
raised by the API responses.

For quite some time, interacting with the API response errors has been
difficult due to them spitting out parsed JSON (as a string). This means
if you need to check the type of exception in order to recover or retry,
you need to use an unreliable `strings.Contains` check. This also
extends to internally when we want to show the user an error, we had a
`errorFromResponse` function which would return the HTTP status code and
whatever was in the `errors` array.

To improve the situation I'm proposing a revamp of our error handling by
exposing various methods on `APIRequestError` struct as helpers to get
the information clients may need to better handle errors. Here are some
examples:

### Return the whole error message

```go
fmt.Println(err.Error())

// => "HTTP status 400: Authentication error (10000)"
```

### Return only the HTTP status

```go
fmt.Println(err.(*cloudflare.APIRequestError).HTTPStatusCode())

// => 400
```

### Check if the response was rate limited

```go
fmt.Println(err.(*cloudflare.APIRequestError).HTTPStatusCode())

// => false
```

In order to achieve this, we'll need to update almost every method that
makes an API call to reflect the following (example from
`createAccessRule` in firewall.go):

```diff
@@ -218,7 +218,7 @@ func (api *API) createAccessRule(prefix string, accessRule AccessRule) (*AccessR
        uri := prefix + "/firewall/access_rules/rules"
        res, err := api.makeRequest("POST", uri, accessRule)
        if err != nil {
-               return nil, errors.Wrap(err, errMakeRequestError)
+               return nil, err.(*APIRequestError)
        }

        response := &AccessRuleResponse{}
```

Personally, I think this is a better situation for the maintainers and
consumers of this library as we'll be better equipped to handle
differing error scenarios. The Terraform provider will also benefit from
these changes for things like retries.