options: Add BaseURL for clients
Updates the client to be able to override the `baseURL`[1] when
instantiating the client.

Use cases are generally restricted to Cloudflare employees but who knows
who else may have use for this.

Here is an example of setting up two clients and compare the output from
local and production and how we would use this.

```go
package main

import (
	"fmt"

	"github.com/cloudflare/cloudflare-go"
)

func main() {
	local, lErr := cloudflare.New(
		"API_KEY",
		"API_EMAIL",
		cloudflare.BaseURL("https://api.test"),
	)

	prod, pErr := cloudflare.New(
		"API_KEY",
		"API_EMAIL",
	)

	if lErr != nil || pErr != nil {
		fmt.Errorf("failed to create clients")
	}

	z1, _ := local.ZoneIDByName("example.com")
	z2, _ := prod.ZoneIDByName("example.com")

	// compare the outputs
}
```

[1]: https://github.com/cloudflare/cloudflare-go/blob/1426b776f85bc6c5c9390247d48b6ea0bcffa13f/cloudflare.go#L22