Deprecate virtualdns.go in favor of dns_firewall.go
This adds new types and functions for the DNS Firewall and DNS Firewall
Analytics API to `dns_firewall.go`, deprecating `virtual_dns.go`

For compatibility, the functions in `virtual_dns.go` are now wrappers
around the functions in `dns_firewall.go`. This means that we now always
use the new routes, even when one of the deprecated functions is called.

In addition to replacing "Virtual DNS" with "DNS Firewall", a few minor
changes were also made to names and function signatures.

Finally, the new DNS Firewall equivalents of the following functions are
no longer exported:

- VirtualDNSResponse
- VirtualDNSListResponse
- VirtualDNSAnalyticsResponse