argo_tunnel: Deprecate `ArgoTunnel` methods in favour of `Tunnel` methods
The Argo Tunnel product has been rebadged to be Tunnels and is due to be
sunset in Feb 2024[1].

This brings the library in line with that expectation and deprecates the
older methods. The API is 1:1 compatible so just the method names are
changing. In addition, I've moved the newer methods to follow the
upcoming experimental changes[2] for parameters which makes the methods
more consistent for use.

As a fallback measure I also updated the ArgoTunnel methods to use the
`cfd_tunnel` URL to ensure we don't have anyone left behind but the plan
is to remove these methods with the API endpoints so best poke people to
move now.

Closes #832

[1]: https://api.cloudflare.com/#argo-tunnel-properties
[2]: https://github.com/cloudflare/cloudflare-go/blob/master/docs/experimental.md