experimental: swap method signature conventions
The original purpose of exposing everything using structs was to be (overly)
explicit about what level of resource we were operating on and allow consistent
customisation between resources. To reduce the verbosity of some of these
structs, we've opted to introduce a `ResourceContainer` concept which accepts
an ID and sets up the method with the same level of information we were
previously using in a more friendly way. Updated recently touched resources to
match this potentially changing the external interface slightly.