cloudflare: update `RequestError` to expose `Messages`
The messages field isn't really used broadly in the API as it was designed for
informational feedback as opposed to errors, however, the page rules endpoint
uses it to communicate validation failures. For these messages to be raised in
other tools, namely the Terraform provider, it needs a dedicated field on the
struct and to be included in the `Error()` method.

Closes cloudflare/terraform-provider-cloudflare#1579