cloudflare: move HTTP request debugging to `httputil`
Within the library, we provide a way to output the HTTP interactions
that the library makes. We previously used a custom approach and format
the output. Even though it works, there are better approaches using
`httputil.DumpRequestOut` and `httputil.DumpResponseOut` for dumping the
HTTP interactions.

Alongside this work, we're adding support for redacting sensitive values
in the HTTP interactions. This is useful for both this library and the
consumers of this library (like the Terraform Provider) to prevent
leaking sensitive information in logs.

Closes #1143