Match ListDNSRecords per_page value to the API's
As outlined in github.com/cloudflare/cloudflare-go/pull/1169,
when no pagination options are passed in to ListDNSRecords,
autopagination is performed. However, the current PerPage
value doesn't match that of the official API. This commit
makes both values match. A new test is also implemented that
more explicitly tests the auto pagination behaviour.

Signed-off-by: Artur Rodrigues <artur.rodrigues@lacework.net>