zone: make settings more flexible
For the life of this library, the settings have always been under the
`/zones/:zone_id/settings` route which was an aggregate of all the
settings. This was designed as a way to aggregate all of the settings in
one place and had its place however, now we are breaking up the zone
settings to their true paths and need to make the `Get` and `Update`
zone settings compatible with this approach.

This will enable teams to use their true paths instead of the prefixed
or aggregate paths with no friction from the library.