Fix: 'create-or-update' handling of 'name' with 'zone' value
In flarectl, the 'flarectl dns create' command can properly process a
'name' parameter even if it includes the value of the 'zone'.
However, the 'create-or-update' command does not work if the 'name'
parameter contains the value of the 'zone'. I have made a change so that
the 'create-or-update' command works even if the 'name' parameter
includes the value of the 'zone'.