split usage of `ID` and `Name`
Pages, unfortunately, uses a friendly name for the identifier in the URL
so to be able to update the `Name` value we need a way to differentiate
the two. The way I've solved this here is that `ID` is used for
addressing the resource via the UI or a stable identifier whereas `Name`
is used for updating the value.