Fix disabling redact PII on logging settings
Trying to disable redact PII on the logging
settings would have no effect because the logging
settings object would be marshaled without the
`redact_pii` field which result in the default
value being used, which is `true` in this case.

To fix this, the `redact_pii` is now included
even set to `false` when the logging settings
object is marshaled.